+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning].
 
 - [PR-92](https://github.com/itk-dev/ai-screening/pull/92)
   - Update drupal core
+- [PR-91](https://github.com/itk-dev/ai-screening/pull/91)
+  - Updated ethics track
 - [PR-89](https://github.com/itk-dev/ai-screening/pull/89)
   - JS filter for law report
   - Changes in report display